// src/redux/user/user.types.ts
export enum UserActionTypes {
  SIGN_OUT = 'user/SIGN_OUT',
  SIGN_UP_START = 'user/SIGN_UP_START',
  SIGN_UP_SUCCESS = 'user/SIGN_UP_SUCCESS',
  SIGN_UP_FAILURE = 'user/SIGN_UP_FAILURE',
  SIGN_IN_START = 'user/SIGN_IN_START',
  SIGN_IN_SUCCESS = 'user/SIGN_IN_SUCCESS',
  SIGN_IN_FAILURE = 'user/SIGN_IN_FAILURE',
  GET_CURRENT_USER_START = 'user/GET_CURRENT_USER_START',
  GET_CURRENT_USER_SUCCESS = 'user/GET_CURRENT_USER_SUCCESS',
  GET_CURRENT_USER_FAILURE = 'user/GET_CURRENT_USER_FAILURE',
  CHANGE_USER_START = 'user/CHANGE_USER_START',
  CHANGE_USER_SUCCESS = 'user/CHANGE_USER_SUCCESS',
  CHANGE_USER_FAILURE = 'user/CHANGE_USER_FAILURE',
  UPDATE_CURRENT_USER = 'user/UPDATE_CURRENT_USER',
  COMPRESS_AND_CHANGE_USER_AVATAR = 'user/COMPRESS_AND_CHANGE_USER_AVATAR',
}